+++ src/cli/index.js
@@ -53,11 +53,7 @@ export function cli(args, cwd, ui = UI) {
         ui.exit(code ? code : 0);
       })
       .catch((error) => {
-        if (error instanceof AzkError) {
-          ui.fail(error.toString());
-        } else {
-          ui.fail(error.stack ? error.stack : error);
-        }
+        ui.fail(error);
         ui.exit(error.code ? error.code : 127);
       });
   }